Fix types and comments

+++ linked list/singly linked list/include/sll_node.h
@@ -4,9 +4,9 @@
 
 typedef struct sll_node
 {
-    int element;
+    size_t element;
     struct sll_node *next;
 } sll_node_t;
 
-sll_node_t *sll_node_alloc(const int element, sll_node_t *next);
+sll_node_t *sll_node_alloc(const size_t element, sll_node_t *next);
 void sll_node_dealloc(sll_node_t **sll_node);
